Solving an Error

+++ Unit_3/Lesson_4_Assignment/Lab_3/startup.c
@@ -12,8 +12,9 @@ void H_fault_Handler() __attribute__((weak, alias("Default_Handler")));;
 
 static uint32 stack_top[256];
 
-void (* const g_p_func_vectors[])() __attribute__((section(".vectors"))) = {
-	(void (*)()) ((uint32)stack_top + sizeof(stack_top)),
+void (* const g_p_func_vectors[])() __attribute__((section(".vectors"))) =
+{
+	(void (*)()) ((unsigned long)stack_top + sizeof(stack_top)),
 	&Reset_Handler,
 	&NMI_Handler,
 	&H_fault_Handler
@@ -27,8 +28,8 @@ extern uint32 _E_BSS;
 
 void Reset_Handler() {
 	uint32 DATA_size = (uint16 *)&_E_DATA - (uint16 *)&_S_DATA;
-	uint16* P_src = (uint16*) & _E_TEXT;
-	uint16* P_dst = (uint16*) & _S_DATA;
+	uint16* P_src = (uint16*)&_E_TEXT;
+	uint16* P_dst = (uint16*)&_S_DATA;
 	for (uint32 i = 0; i < DATA_size; i++) {
 		*((uint16*)P_dst++) = *((uint16*)P_src++);
 	}
